Extract the dirties from the ViewModel
